python/Sashimi/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name='sashimi'
version='3.0.0'
description='3D printer stacking software'
requires-python=">=3.11,<3.12"
authors=[{name='Ross Marchant', email='ross.g.marchant@gmail.com'},
         {name='Louis Lanoy', email='louis.lanoy@icloud.com'}]
dependencies=[
        'numpy==1.26.4',
        'scikit-image==0.22.0',
        'scipy==1.12.0',
        'pypylon==3.0.1',
        'opencv-python==4.9.0.80',
        'pyserial==3.5',
        'click==8.1.7',
        'pillow==10.2.0',
        'PySide6==6.6.1',
        'marshmallow_dataclass==8.6.0',
        'tqdm',
        'itk-montage'
    ]